chore: 클래스 분리하여 파일 구조 변경
