+++ lib/asset.dart
@@ -1,0 +1,6 @@
+// 가 : 자산
+abstract class Asset{
+  String name;
+
+  Asset({required this.name});
+}
