+++ lib/tangible_asset.dart
@@ -1,35 +1,17 @@
-// 가 : 자산
-abstract class Asset{
-  String name;
-
-  Asset({required this.name});
-}
-
-// 형태가 있는것
-abstract interface class Thing{
-  double get weight;
-  set weight(value);
-}
-
-// 나 : 무형자산
-abstract class IntangibleAsset {}
-
-// 다 : 특허권
-class Patent extends IntangibleAsset{}
-
-abstract class TangibleAsset extends Asset implements Thing {
-  int price;
-  String color;
-
-  TangibleAsset({required super.name, required this.price, required this.color, });
-}
+import 'thing.dart';
 
 class Book extends TangibleAsset{
   String isbn;
   double _weight;
 
   // super named
-  Book({required super.name, required this.isbn, double weight = 0, required super.price, required super.color,}) : _weight = weight{
+  Book({
+    required super.name, 
+    required this.isbn, 
+    double weight = 0, 
+    required super.price, 
+    required super.color,
+  }) : _weight = weight {
     this.weight = _weight;
   }
 
@@ -40,7 +22,6 @@ class Book extends TangibleAsset{
   }
   
   @override
-  // TODO: implement weight
   double get weight => _weight;
 }
 
@@ -48,7 +29,13 @@ class Computer extends TangibleAsset{
   String makerName;
   double _weight;
 
-  Computer({required super.price, required super.color, required super.name, double weight = 0, required this.makerName}) : _weight = weight{
+  Computer({
+    required super.price, 
+    required super.color, 
+    required super.name, 
+    double weight = 0, 
+    required this.makerName
+  }) : _weight = weight {
     this.weight = _weight;
   }
